refactor: callbacks + metrics modules
