+++ trackers/core/reid/trainers/callbacks.py
